* feature/plant-details:
  Added flower flip event
  Added event deleting
  Added action for actionlist
  Added basic action adapter functionality
  Added basic event fragment/adapter
  Added add photo button style
  Added add photo fab functionality
  Added zoomed image functionality
  Added basic view photos activity
  Save when action/image is made and display last image in plant list
  Added UIL
  Added basic add photo functionality
  Added action and undo
  Added nutrient popup and more nutrient fields
  Fixed issue with actions not saving correctly
  Added basic add feed functionality
  Added basic add feeding activity/fragment
  Added action/links container visible for edit
  Added options and links
  Fixed issues with lombok, gson and passing objects around activities

+++ app/src/main/java/me/anon/model/Feed.java
@@ -15,5 +15,5 @@ import lombok.experimental.Accessors;
 public class Feed extends Water
 {
 	private Nutrient nutrient;
-	private double mlpl; // ml per litre
+	private int mlpl; // ml per litre
 }
